requirements.txt
# meta requirements
virtualenv
pre-commit
# external requirements
flake8==3.8.3                       # MIT
jupyter==1.0.0                      # BSD
jupyter_contrib_nbextensions==0.5.1 # BSD
ipykernel==5.3.2                    # BSD
black==19.10b0                      # MIT
seaborn==0.10.1                     # BSD 3-clause
# Install pytorch as per your systems configguration

easydict
comet-ml
pytorch-lightning
scikit-learn
pytorch-lightning-bolts==0.2.2
